# fix EachKey pIdxFlags allocation